Update patch set 9
Patch Set 9:

Thanks for the suggestions. As to the abbreviation, I do have some discussions offline. It would be better you suggest some candidates. For the system bandwidth test case, our purpose is to profile an untested system or the system that is not properly configured or provisioned. Then the results could be analyzed and this bottleneck will be decoupled into different categories, e.g., compute, memory, network, mid-ware, APP. In the end, some algorithm for auto staging will be developed/used to tune the parameters, thanks!

Patch-set: 9