Update test suite running entrance
JIRA: BOTTLENECK-142

Update the entrance for POSCA test suite.Currently, test case
selection for POSCA is not support for this entrance.
The function will be added later.
For Rubbos and VSTF, after OPNFV Colorado release,
they are not updating anymore. This entrance for running them
within Bottlenecks is no longer supported.
These testsuites are also not in the release plan with Bottlenecks
since then. If you want to run Rubbos and VSTF,
please refer to ealier releases.

Change-Id: Idaf7da147489c6371e46ac3234179133762a05cd
Signed-off-by: yuyang <Gabriel.yuyang@huawei.com>