Update patch set 3
Patch Set 3:

(1 comment)

I've currently placed this script in the /bottlenecks/monitor/ directory on the host. I can copy the same to the docker container in a directory like /etc/grafana and give path to the dashboard in the container and run this script there as well. So currently it is on the host.

Patch-set: 3