Remove inactive committers
The TSC decided to remove the committers that have been inactive in gerit
for a year or more. If any of the people removed by this patch are contributing
to the project in other ways, and the PTL would like to keep the gerrit rights
for them, please ammend this patch.

For other updates on the committer list, please follow the procedure descrbed in
https://wiki.opnfv.org/display/DEV/Developer+governance+and+reports

JIRA: TSC-14

Change-Id: I5eda05afbc087034e5a399632749ca1a1b3fedae
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>